+++ spec/avro/dsl/test/recursive.rb
@@ -1,0 +1,4 @@
+record :recursive, namespace: :test do
+  required :s, :string
+  optional :child, :recursive
+end